README.md
In this video, I walk you through the process of creating a ticket on ServiceNow. I start by signing up using a temporary email and then proceed to create an instance in Yokohama. Once I'm in, I navigate to the incidents section and prepare to create a new ticket for a password account lockout issue. I encourage you to follow along and replicate the steps as I demonstrate them. Let’s get started!